cleanup comments etc

+++ src/loki/transit_available_action.cc
@@ -29,8 +29,7 @@ std::string loki_worker_t::transit_available(Api& request) {
       const auto& ll = location.latlng_;
       DistanceApproximator<PointLL> approximator(ll);
       double latdeg = double(location.radius_) / kMetersPerDegreeLat;
-      double lngdeg =
-          double(location.radius_) / DistanceApproximator<PointLL>::MetersPerLngDegree(ll.lat());
+      double lngdeg = location.radius_ / DistanceApproximator<PointLL>::MetersPerLngDegree(ll.lat());
       AABB2<PointLL> bbox(Point2d(ll.lng() - lngdeg, ll.lat() - latdeg),
                           Point2d(ll.lng() + lngdeg, ll.lat() + latdeg));
       std::vector<int32_t> tilelist = tiles.TileList(bbox);
